backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0
Pillow>=10.1,<11.0
celery>=5.3,<6.0
redis>=5.0,<6.0
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0
drf-spectacular>=0.27,<1.0
dj-database-url>=2.1,<3.0
